Create new const file
